REFACTORING Servlet to Controller
